Переместил проект из основной папки в модуль task2
