+++ Homework/build.gradle
@@ -27,17 +27,10 @@ dependencies {
 	implementation project(':task2')
 
 	implementation 'org.springframework.boot:spring-boot-starter'
-
-	// Jackson
-	implementation 'com.fasterxml.jackson.core:jackson-databind:2.17.2'
-	implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-xml:2.15.0'
-
-	// Lombok
 	compileOnly 'org.projectlombok:lombok'
-	annotationProcessor 'org.projectlombok:lombok'
-
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
 	annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
+	annotationProcessor 'org.projectlombok:lombok'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 }
