+++ Homework/task2/src/main/java/org/yakdanol/task2/model/City.java
@@ -1,4 +1,4 @@
-package org.yakdanol.homework.model;
+package org.yakdanol.task2.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
