+++ Homework/task2/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=task2
+spring.application.name=Homework
